Minor Tweaks and Implemented a Pseudo PID
I changed some parameters to make this more effective; we can test for perfection obviously. I added a "Pseudo PID" where I am basically dividing the distance from hatch by a constant (in this case 20). This allows for the robot to slow down as it gets closer and closer to the hatch. FInally I added a command group where we can run 3 commands in order. This should allow us to first go forward second close the clapper to release the hatch and 3rd go backwards